refactor(click): extract handle click
