+++ index.js
@@ -1,32 +1,7 @@
 'use strict';
 
 const isFirefoxAndroid = require('./lib/is-firefox-android')();
-const tabs = require('sdk/tabs');
-
-function handleClick() {
-  const worker = tabs.activeTab.attach({
-    contentScriptFile: './content-script.js',
-  });
-  worker.port.on('copyToSystem', (request) => {
-    if (isFirefoxAndroid) {
-      const { Cc, Ci } = require('chrome');
-      const prompts = Cc['@mozilla.org/embedcomp/prompt-service;1']
-        .getService(Ci.nsIPromptService);
-      prompts.prompt(
-        null,
-        'Build Link Plain',
-        'Built link',
-        { value: request },
-        null,
-        { value: false }
-      );
-    } else {
-      const clipboard = require('sdk/clipboard');
-      clipboard.set(request);
-    }
-  });
-  worker.port.emit('buildLinkPlain');
-}
+const handleClick = require('./lib/handle-click');
 
 if (isFirefoxAndroid) {
   const getWindow = require('./lib/get-window');
